Move everything to colour variables
