+++ .coderabbit.yml
@@ -3,5 +3,5 @@ reviews:
   collapse_walkthrough: true
   sequence_diagrams: false
   path_filters:
-    - "!src/components/ui"
-    - "!src/hooks"
+    - "!src/components/ui/**"
+    - "!src/hooks/**"
