+++ .devcontainer/devcontainer.json
@@ -29,7 +29,9 @@
         "bradlc.vscode-tailwindcss",
         "csstools.postcss",
         "mongodb.mongodb-vscode",
-        "vk-l.nanoid-generator"
+        "vk-l.nanoid-generator",
+        "nize.oklch-preview",
+        "starbist.oklchanger"
       ]
     }
   },
