+++ .vscode/extensions.json
@@ -8,6 +8,8 @@
     "bradlc.vscode-tailwindcss",
     "csstools.postcss",
     "mongodb.mongodb-vscode",
-    "vk-l.nanoid-generator"
+    "vk-l.nanoid-generator",
+    "nize.oklch-preview",
+    "starbist.oklchanger"
   ]
 }
