+++ src/components/answer/answer.css
@@ -1,0 +1,17 @@
+@import "tailwindcss";
+
+.answer-container {
+  @apply w-[800px] bg-[var(--muted)];
+}
+
+.valid-plan {
+  @apply flex items-center gap-2;
+}
+
+.valid-icon {
+  @apply w-5 h-5 text-[var(--color-green-600)];
+}
+
+.invalid-icon {
+  @apply w-5 h-5 text-[var(--color-red-600)];
+}
