+++ src/components/app-sidebar/scenarioItem.tsx
@@ -1,17 +1,14 @@
 import { ScenarioData } from "@/models/scenario";
 import Link from "next/link";
 import { useEffect, useRef } from "react";
-import { SidebarMenuButton, SidebarMenuItem } from "./ui/sidebar";
+import { SidebarMenuButton, SidebarMenuItem } from "@/components/ui/sidebar";
 
 interface ScenarioItemProps {
   scenario: ScenarioData;
   selected: boolean;
 }
 
-export default function ScenarioItem({
-  scenario,
-  selected,
-}: ScenarioItemProps) {
+export function ScenarioItem({ scenario, selected }: ScenarioItemProps) {
   const { plan } = scenario;
 
   const ref = useRef<HTMLLIElement>(null);
