--- src/styles/answer.css
@@ -1,22 +1,0 @@
-.answer-container {
-  width: 800px;
-  background: var(--muted);
-}
-
-.valid-plan {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.valid-icon {
-  width: 20px;
-  height: 20px;
-  color: #16a34a; /* text-green-600 */
-}
-
-.invalid-icon {
-  width: 20px;
-  height: 20px;
-  color: #dc2626; /* text-red-600 */
-}
