--- src/styles/conversation.css
@@ -1,43 +1,0 @@
-.chat-row {
-  display: flex;
-  margin: 0.75em 0;
-}
-
-.chat-row.left {
-  justify-content: flex-start;
-}
-
-.chat-row.right {
-  justify-content: flex-end;
-}
-
-.chat-bubble {
-  max-width: 70%;
-  font-family: sans-serif;
-}
-
-.label {
-  font-size: 0.75em;
-  font-weight: bold;
-  margin-bottom: 0.25em;
-}
-
-.bubble {
-  padding: 0.75em 1em;
-  border-radius: 1em;
-  line-height: 1.4;
-  word-wrap: break-word;
-  white-space: pre-line;
-}
-
-.pilot .bubble {
-  background-color: #e0f7fa;
-  color: #006064;
-  border-top-left-radius: 0;
-}
-
-.controller .bubble {
-  background-color: #f1f8e9;
-  color: #33691e;
-  border-top-right-radius: 0;
-}
