--- src/styles/sidebar.css
@@ -1,14 +1,0 @@
-/* styles/layout.css */
-@import "tailwindcss";
-
-.sidebar {
-  @apply bg-gray-100 border-r border-gray-300 p-4 h-screen overflow-y-auto;
-}
-
-.sidebar .header {
-  @apply bg-gray-100 pt-3 pb-3 font-semibold text-lg;
-}
-
-.sidebar ul {
-  @apply space-y-2;
-}
